pyproject.toml
[project]
name = "funtech-test"
version = "0.1.0"
description = "123"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "fastapi-limiter>=0.1.6",
    "fastapi>=0.115.12",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.8.1",
    "pydantic[email]>=2.11.1",
    "python-jose>=3.4.0",
    "python-multipart>=0.0.20",
    "redis>=5.2.1",
    "sqlalchemy>=2.0.40",
    "uvicorn>=0.34.0",
    "aio-pika>=9.5.5",
    "celery[amqp]>=5.5.0",
]

[dependency-groups]
dev = [
    "ruff>=0.11.2",
]

[tool.pyright]
venvPath = "."
venv = ".venv"
